Sending Map between Server and Client

+++ Project File/Client and Server Projects/Server/Assets/GenerationScriptV2.cs
@@ -48,8 +48,8 @@ public class GenerationScriptV2 : MonoBehaviour
 
     string worldName;
 
-    int[,] map;
-    int[,] cavemap;
+    byte[,] map;
+    byte[,] cavemap;
     
 
     void Start()
@@ -109,7 +109,7 @@ public class GenerationScriptV2 : MonoBehaviour
         //Debug.Log($"buildblock called, position {x}, {y}");
         if ((map[x-1,y] != 0 || map[x + 1, y] != 0 || map[x, y - 1] != 0 || map[x, y + 1] != 0) && (map[x,y] == 0 || map[x, y] == 8 || map[x, y] == 9))
         {
-            map[x, y] = block;
+            map[x, y] = (byte)block;
             TileBase placing = Plank;
             switch (block) 
             {
@@ -156,7 +156,7 @@ public class GenerationScriptV2 : MonoBehaviour
         StoneSeed = StoneSlider.value;
         //CaveSeed = CaveSlider.value;
         
-        map = new int[worldWidth, worldHeight];
+        map = new byte[worldWidth, worldHeight];
         OptimisedTerrainGeneration(map, worldWidth, worldHeight, GrassSoil, Stone);
         ApplyCaves(map);
         AddTrees(TreePopulation, TestTileFG, Log, Leaf);
@@ -173,7 +173,7 @@ public class GenerationScriptV2 : MonoBehaviour
         //CaveSeed = (float)UnityEngine.Random.Range(0.02f, 0.05f);
         OreSeed = (float)UnityEngine.Random.Range(0.03f, 0.05f);
 
-        map = new int[worldWidth, worldHeight];
+        map = new byte[worldWidth, worldHeight];
 
         if (IsFromSave())
         {
@@ -201,7 +201,7 @@ public class GenerationScriptV2 : MonoBehaviour
         //Renderer(map, GrassSoil, Stone, TestTileMap);
     }
 
-    public int[,] SendMap()
+    public byte[,] SendMap()
     {
         return map;
     }
@@ -279,7 +279,7 @@ public class GenerationScriptV2 : MonoBehaviour
     //    return map;
     //}
 
-    public void Renderer(int[,] MapToRender, Tilemap FG, Tilemap BG)
+    public void Renderer(byte[,] MapToRender, Tilemap FG, Tilemap BG)
     {
         FG.ClearAllTiles();
         BG.ClearAllTiles();
@@ -374,7 +374,7 @@ public class GenerationScriptV2 : MonoBehaviour
                 Debug.Log(Convert.ToInt16(value[1]));
                 for (int x = 0; x < map.GetLength(0); x++)
                 {
-                    map[x, y] = Convert.ToInt16(value[x]) - 48;
+                    map[x, y] = (byte)(Convert.ToInt16(value[x]) - 48);
                 }
                 y++;
             }
@@ -390,7 +390,7 @@ public class GenerationScriptV2 : MonoBehaviour
     /// <param name="GrassSoil"></param>
     /// <param name="Stone"></param>
     /// <param name="TestTileMap"></param>
-    public void OptimisedTerrainGeneration(int[,] WorldMap, int width, int height, TileBase GrassSoil, TileBase Stone)
+    public void OptimisedTerrainGeneration(byte[,] WorldMap, int width, int height, TileBase GrassSoil, TileBase Stone)
     {
         int perlinNoiseSoil;
         int perlinNoiseStone;
@@ -436,7 +436,7 @@ public class GenerationScriptV2 : MonoBehaviour
     /// <param name="map"></param>
     /// <param name="width"></param>
     /// <param name="height"></param>
-    public void ApplyCaves(int[,] map)
+    public void ApplyCaves(byte[,] map)
     {
         //Version 1
         //for (int x = 0; x < width; x++)
@@ -481,12 +481,12 @@ public class GenerationScriptV2 : MonoBehaviour
 
     public void CaveGeneration()
     {
-        cavemap = new int[worldWidth, worldHeight];
+        cavemap = new byte[worldWidth, worldHeight];
         for (int x = 0; x < worldWidth; x++)
         {
             for (int y = 0; y < worldHeight; y++)
             {
-                cavemap[x, y] = UnityEngine.Random.Range(0, 100) < RandomPercentFill ? 1 : 0;
+                cavemap[x, y] = (byte)(UnityEngine.Random.Range(0, 100) < RandomPercentFill ? 1 : 0);
             }
         }
     }
@@ -495,7 +495,7 @@ public class GenerationScriptV2 : MonoBehaviour
     {
         for (int i = 1; i <= count; i++)
         {
-            int[,] tempmap = (int[,])cavemap.Clone();
+            byte[,] tempmap = (byte[,])cavemap.Clone();
             //Debug.Log(tempmap.GetLength(0));
             //Debug.Log(tempmap.GetLength(1));
 
@@ -520,7 +520,7 @@ public class GenerationScriptV2 : MonoBehaviour
 
 
 
-    private int GetNeighbours(int[,] tempmap, int xs, int ys, int neighbours)
+    private int GetNeighbours(byte[,] tempmap, int xs, int ys, int neighbours)
     {
         for (int x = xs - 1; x <= xs + 1; x++)
         {
